+++ Doc/source/release_notes.rst
@@ -21,8 +21,9 @@ New features
 pandas 0.18.0, which is not installed by default with SpacePy.
 
 `~.datamodel.dmarray` supports output to Astropy Quantities
-(`~.datamodel.ISTPArray.toQuantity`). This support requires
-Astropy 1.0, which is not installed by default with SpacePy.
+(`~.datamodel.ISTPArray.toQuantity`) and input from them
+(`~.datamodel.ISTPArray.fromQuantity`). This support requires Astropy
+1.0, which is not installed by default with SpacePy.
 
 0.5 Series
 ==========